requirements.txt
apache-airflow==2.7.2
boto3
pandas
pyarrow
requests